seeds fixed

+++ db/seeds.rb
@@ -6,7 +6,13 @@
 #   movies = Movie.create([{ name: "Star Wars" }, { name: "Lord of the Rings" }])
 #   Character.create(name: "Luke", movie: movies.first)
 
-Medication.destroy_all && User.destroy_all
+Intake.destroy_all && Medication.destroy_all && User.destroy_all
+
+interval = ["Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"]
+intake_times = [
+  "1:00am", "2:00am", "3:00am", "4:00am", "5:00am", "6:00am", "7:00am", "8:00am", "9:00am", "10:00am", "11:00am",
+  "1:00pm", "2:00pm", "3:00pm", "4:00pm", "5:00pm", "6:00pm", "7:00pm", "8:00pm", "9:00pm", "10:00pm", "11:00pm"
+]
 
 ms_von_braun = User.create!(
   email: "von_braun@gmail.com",
@@ -97,3 +103,15 @@ naproxen = Medication.create!(
   dose: "2 Tab",
   user: ms_von_braun
 )
+
+aspirin.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+ibuprofen.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+multivitamin.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+magnesium.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+calcium.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+vitamind.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+vitaminb.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+antibiotika.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+vitaminc.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+diclofenac.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
+naproxen.generate_intakes(interval.sample(rand(1..3)), intake_times.sample(rand(1..3)))
